requirements.txt
openai==1.12.0
python-dotenv==1.0.1
transformers==4.36.0
torch==2.2.0
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
tqdm==4.66.1
streamlit==1.32.0
plotly==5.18.0
accelerate==0.27.0
einops==0.7.0
sentencepiece==0.2.0 